+++ src/Phan/Language/Element/VariadicParameter.php
@@ -103,7 +103,7 @@ class VariadicParameter extends Parameter
         // e.g. $this->getUnionType() is of type T[]
         //      $this->non_variadic->getUnionType() is of type T
         return new Parameter(
-            // @phan-suppress-next-line PhanTypeMismatchArgument Here it's fine to pass a FileRef
+            // @phan-suppress-next-line PhanTypeMismatchArgumentSuperType Here it's fine to pass a FileRef
             $this->getFileRef(),
             $this->getName(),
             $this->type,